requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt --python-version 3.12
alembic==1.16.5
    # via -r requirements.in
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   starlette
    #   watchfiles
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.2.0
    # via streamlit
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   streamlit
    #   uvicorn
fastapi==0.116.1
    # via -r requirements.in
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
h11==0.16.0
    # via uvicorn
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
narwhals==2.3.0
    # via altair
numpy==2.3.2
    # via
    #   pandas
    #   pydeck
    #   streamlit
packaging==25.0
    # via
    #   altair
    #   streamlit
pandas==2.3.2
    # via streamlit
pillow==11.3.0
    # via streamlit
protobuf==6.32.0
    # via streamlit
psycopg2-binary==2.9.10
    # via -r requirements.in
pyarrow==21.0.0
    # via streamlit
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.33.2
    # via pydantic
pydeck==0.9.1
    # via streamlit
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   uvicorn
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via uvicorn
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via streamlit
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via
    #   -r requirements.in
    #   alembic
starlette==0.47.3
    # via fastapi
streamlit==1.49.1
    # via -r requirements.in
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.2
    # via streamlit
typing-extensions==4.15.0
    # via
    #   alembic
    #   altair
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   starlette
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
